+++ frontend/src/api/services/image.js
@@ -1,12 +1,12 @@
-import axiosInstance from '../axiosInstance';
-import { ENDPOINTS } from '../endpoints';
+import axiosInstance from '@api/axiosInstance';
+import { ENDPOINTS } from '@api/endpoints';
 
 /**
  * Uploads a new profile image to the server as multipart/form-data.
  */
 export async function uploadProfileImage(formData) {
   const { data } = await axiosInstance.post(ENDPOINTS.image.profile, formData, {
-    headers: { 'Content-Type': 'multipart/form-data' },
+    headers: { 'Content-Type': 'multipart/form-data' }, // TODO: This is wrong, look at backend test how it's sent
   });
   return data;
 }
